Move all release UI components to components folder
