+++ static/js/publisher/release/components/devmodeIcon.js
@@ -1,9 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-export function isInDevmode(revision) {
-  return revision.confinement === "devmode" || revision.grade === "devel";
-}
+import { isInDevmode } from "../helpers";
 
 export default function DevmodeIcon({ revision, showTooltip }) {
   return (
