+++ static/js/publisher/release/components/historyPanel.js
