+++ static/js/publisher/release/components/notification.js
