+++ static/js/publisher/release/helpers.js
@@ -1,0 +1,3 @@
+export function isInDevmode(revision) {
+  return revision.confinement === "devmode" || revision.grade === "devel";
+}
